test(bigquery): test all function for bigquery
