+++ ibis-server/tests/routers/v3/connector/bigquery/test_functions.py
@@ -1,4 +1,5 @@
 import base64
+import os
 
 import orjson
 import pytest
@@ -8,6 +9,9 @@ from app.config import get_config
 from app.main import app
 from tests.conftest import DATAFUSION_FUNCTION_COUNT
 from tests.routers.v3.connector.bigquery.conftest import base_url, function_list_path
+from tests.util import FunctionCsvParser, SqlTestGenerator
+
+pytestmark = pytest.mark.functions
 
 manifest = {
     "catalog": "my_catalog",
@@ -47,15 +51,15 @@ with TestClient(app) as client:
         response = client.get(url=f"{base_url}/functions")
         assert response.status_code == 200
         result = response.json()
-        assert len(result) == DATAFUSION_FUNCTION_COUNT + 34
-        the_func = next(filter(lambda x: x["name"] == "abs", result))
+        assert len(result) == DATAFUSION_FUNCTION_COUNT + 22
+        the_func = next(filter(lambda x: x["name"] == "string_agg", result))
         assert the_func == {
-            "name": "abs",
-            "description": "Returns the absolute value of a number.",
-            "function_type": "scalar",
+            "name": "string_agg",
+            "description": "Aggregates string values with a delimiter.",
+            "function_type": "aggregate",
             "param_names": None,
-            "param_types": None,
-            "return_type": "FLOAT64",
+            "param_types": "text,text",
+            "return_type": "text",
         }
 
         config.set_remote_function_list_path(None)
@@ -97,3 +101,18 @@ with TestClient(app) as client:
             "data": [[1]],
             "dtypes": {"col": "int64"},
         }
+
+    def test_functions(manifest_str: str, connection_info):
+        csv_parser = FunctionCsvParser(os.path.join(function_list_path, "bigquery.csv"))
+        sql_generator = SqlTestGenerator("bigquery")
+        for function in csv_parser.parse():
+            sql = sql_generator.generate_sql(function)
+            response = client.post(
+                url=f"{base_url}/query",
+                json={
+                    "connectionInfo": connection_info,
+                    "manifestStr": manifest_str,
+                    "sql": sql,
+                },
+            )
+            assert response.status_code == 200
